requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cookiecutter==2.6.0
Django==4.2.11
django-modeltranslation==0.18.11
django-tailwind==3.6.0
exceptiongroup==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.0.3
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
locust==2.24.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
psutil==5.9.8
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.17.2
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
zope.event==5.0
zope.interface==6.2
